refactor of layout
