+++ backend/db/database.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from settings import SQLALCHEMY_DATABASE_DIR
+from backend.settings import SQLALCHEMY_DATABASE_DIR
 
 engine = create_engine(
     SQLALCHEMY_DATABASE_DIR, connect_args={"check_same_thread": False}
