+++ backend/db/models.py
@@ -1,4 +1,4 @@
-from db.database import Base
+from backend.db.database import Base
 from sqlalchemy import Column, DateTime
 from sqlalchemy.sql.sqltypes import Integer, String
