+++ frontend/Dockerfile
@@ -1,8 +1,8 @@
 FROM python:3.11.2
 WORKDIR /app
-COPY requirements.txt ./requirements.txt
+COPY ../requirements.txt ./requirements.txt
 RUN pip install -r requirements.txt
 EXPOSE 8501
-COPY . /app
+COPY .. /app
 ENTRYPOINT [ "streamlit", "run" ]
 CMD ["streamlit_app.py"]
